warehouse_project/warehouse/warehouse/settings/prod_site_en.py
from .base import *
from dotenv import load_dotenv
import os

load_dotenv()

DEBUG = False

SECRET_KEY = os.getenv('SECRET_KEY')

ALLOWED_HOSTS = ['*']

SITE_ID = 3

STATIC_ROOT = BASE_DIR / '..' / 'nginx' / 'static'

DATABASES = {
    'default': {
        'ENGINE': 'django.db.backends.postgresql',
        'NAME': os.getenv('DB_NAME'),
        'USER': os.getenv('DB_USERNAME'),
        'PASSWORD': os.getenv('DB_PASSWORD'),
        'HOST': os.getenv('DB_HOST'),
        'PORT': os.getenv('DB_PORT'),
    }
}
