warehouse_project/warehouse/requirements.txt
django==3.2.9
pillow
python-dotenv
psycopg2-binary
